Extract openings list components
